initial cleanup

+++ includes/999-cleanup.sh
@@ -1,4 +1,5 @@
-# Copyright 2013, Joyent. Inc. All rights reserved.
+#!/usr/bin/env bash
+# Copyright (c) 2017, Joyent. Inc.
 
 log "cleaning up"
 
@@ -6,5 +7,12 @@ svccfg -s zoneinit 'setprop application/done = true'
 svcadm refresh zoneinit
 rm -f ${ZONECONFIG}
 
-log "scheduling an immediate reboot"
-echo "reboot >/dev/null" | at now >/dev/null
+shouldreboot=$(json -f "$ZONEINIT_DIR/zoneinit.json" features.reboot)
+
+if [[ $shouldreboot == true ]] || (($? != 0)); then
+	# Reboot the zone if features.reboot is true, or if the above call
+	# fails to json.  This way, we are backwards-compatible with versions
+	# of zoneinit that do not have zoneinit.json
+	log "scheduling an immediate reboot"
+	echo "reboot >/dev/null" | at now >/dev/null
+fi
